Use sortingType from state. WIP
WIP: Places component doesnt rerender on sortingType change
